requirements.txt
streamlit==1.45.1
tensorflow==2.19.0
Pillow==11.2.1
numpy==2.1.3
